+++ src/tx-params/convert-sell-all.js
@@ -5,15 +5,14 @@ import {convertToPip} from 'minterjs-util';
 
 /**
  * @constructor
- * @param {string} privateKey
  * @param {string} coinFrom
  * @param {string} coinTo
  * @param {number|string} [minBuyAmount=0]
  * @param {string} [feeCoinSymbol]
- * @param {string} [message]
+ * @param {...TxParams} otherParams
  * @return {TxParams}
  */
-export default function SellAllTxParams({privateKey, coinFrom, coinTo, minBuyAmount = 0, feeCoinSymbol, message}) {
+export default function SellAllTxParams({coinFrom, coinTo, minBuyAmount = 0, feeCoinSymbol, ...otherParams}) {
     const txData = new MinterSellAllTxData({
         coinToSell: formatCoin(coinFrom),
         coinToBuy: formatCoin(coinTo),
@@ -25,8 +24,7 @@ export default function SellAllTxParams({privateKey, coinFrom, coinTo, minBuyAmo
     }
 
     return {
-        privateKey,
-        message,
+        ...otherParams,
         gasCoin: feeCoinSymbol,
         txType: TX_TYPE_SELL_ALL_COIN,
         txData: txData.serialize(),